% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{DOWY_to_Date}
\alias{DOWY_to_Date}
\title{DOWY_to_Date
Takes a day of water year and optionally a year, and returns a date}
\usage{
DOWY_to_Date(DOWY, year = 2020)
}
\arguments{
\item{year}{year, defaults to 2020, but having the correct year will account for leap years}

\item{DOY}{integer day of water year}
}
\value{
a date, formatted as 'ymd'
}
\description{
DOWY_to_Date
Takes a day of water year and optionally a year, and returns a date
}
