% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{NDVI_quantiles}
\alias{NDVI_quantiles}
\title{NDVI_quantiles}
\usage{
NDVI_quantiles(NDVI_Stack, probs = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1))
}
\arguments{
\item{NDVI_Stack}{a raster brick containing the 12 MODIS NDVI metrics}

\item{probs}{a vector of quantile values to calculate}
}
\value{
a dataframe with quantiles of each metric
}
\description{
NDVI_quantiles
}
