% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{NDVI_Summaries}
\alias{NDVI_Summaries}
\title{NDVI_Summaries}
\usage{
NDVI_Summaries(NDVI_Stack)
}
\arguments{
\item{NDVI_Stack}{a raster brick containing the 12 MODIS NDVI metrics}
}
\value{
a list containing table, a dataframe of min, max, mean, and sd each layer in brick, and a list of histograms
}
\description{
NDVI_Summaries
}
