% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{QC_mask}
\alias{QC_mask}
\title{Mask layers of a raster by a specified layer of same raster}
\usage{
QC_mask(x, QC_band = 12, QC_good_value = 1)
}
\arguments{
\item{x}{a terra spatraster a QC layer and other layers}

\item{QC_band}{the layer number with QC data}

\item{QC_good_value}{the value of QC band that indicates good data in other layers}
}
\value{
a spatraster with all but good values masked
}
\description{
Mask layers of a raster by a specified layer of same raster
}
\details{
QC_mask
}
