% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Raster_quantiles}
\alias{Raster_quantiles}
\title{Calculate per-layer quantiles}
\usage{
Raster_quantiles(
  x,
  probs = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1),
  samp_size = 10000
)
}
\arguments{
\item{probs}{a vector of quantile values to calculate}

\item{Raster_Stack}{a raster brick containing the 12 MODIS NDVI metrics}
}
\value{
a dataframe with quantiles of each metric
}
\description{
Generates quantiles for each layer of a raster stack, per-scene, not per-pixel.
}
\details{
Raster_quantiles
}
