% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Raster_Summaries}
\alias{Raster_Summaries}
\title{Calculate per-layer raster summaries}
\usage{
Raster_Summaries(Raster_Stack)
}
\arguments{
\item{Raster_Stack}{a raster brick containing the 12 MODIS NDVI metrics}
}
\value{
a list containing table, a dataframe of min, max, mean, and sd each layer in brick, and a list of histograms
}
\description{
Generates basic summaries for each layer of a raster stack. Calculates minimum, maximum, mean, standard deviation, and histograms, by scene, not by pixel.
}
\details{
Raster_Summaries
}
