% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Trans_rast_list}
\alias{Trans_rast_list}
\title{translate a list of rasters}
\usage{
Trans_rast_list(input_list, layer_names)
}
\arguments{
\item{input_list}{a list of multi-layer spatrasters with the same layer structure}

\item{band_names}{a vector of the names of the layers shared by each input file, will become the names of each element of the top-level list}
}
\value{
a list of lists
}
\description{
takes a list of multilayer spatrasters, and returns a list of multi-layer rasters. The top-level being a list of layers, and each element of that list being a list of that layer from each input file. Useful for taking a multi-year list of multi-band rasters, and returning a multi-band list of multi-year rasters.
}
