% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{DOSY_to_Date}
\alias{DOSY_to_Date}
\title{Day of snow year to date}
\usage{
DOSY_to_Date(DOSY, year = 2020)
}
\arguments{
\item{DOSY}{integer day of snow year}

\item{year}{year, defaults to 2020, but having the correct year will account for leap years}
}
\value{
a date, formatted as 'ymd'
}
\description{
Takes a day of snow year and optionally a year, and returns a date
}
