% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{DOY_to_Date}
\alias{DOY_to_Date}
\title{DOY_to_Date
Takes a day of year and optionally a year, and returns a date}
\usage{
DOY_to_Date(DOY, year = 2020)
}
\arguments{
\item{DOY}{integer day of year}

\item{year}{year, defaults to 2020, but having the correct year will account for leap years}
}
\value{
a date, formatted as 'ymd'
}
\description{
DOY_to_Date
Takes a day of year and optionally a year, and returns a date
}
